don't cast time macros to u32

+++ gc/ogc/lwp_watchdog.h
@@ -29,14 +29,14 @@
 #define TB_NSPERUS					1000
 #define TB_USPERTICK				10000
 
-#define ticks_to_cycles(ticks)		((u32)(((u64)(ticks)*(u64)((TB_CORE_CLOCK*2)/TB_TIMER_CLOCK))/2))
-#define ticks_to_secs(ticks)		((u32)((u64)(ticks)/(u64)(TB_TIMER_CLOCK*1000)))
-#define ticks_to_millisecs(ticks)	((u32)((u64)(ticks)/(u64)(TB_TIMER_CLOCK)))
-#define ticks_to_microsecs(ticks)	((u32)(((u64)(ticks)*8)/(u64)(TB_TIMER_CLOCK/125)))
-#define ticks_to_nanosecs(ticks)	((u32)(((u64)(ticks)*8000)/(u64)(TB_TIMER_CLOCK/125)))
+#define ticks_to_cycles(ticks)		((((u64)(ticks)*(u64)((TB_CORE_CLOCK*2)/TB_TIMER_CLOCK))/2))
+#define ticks_to_secs(ticks)		(((u64)(ticks)/(u64)(TB_TIMER_CLOCK*1000)))
+#define ticks_to_millisecs(ticks)	(((u64)(ticks)/(u64)(TB_TIMER_CLOCK)))
+#define ticks_to_microsecs(ticks)	((((u64)(ticks)*8)/(u64)(TB_TIMER_CLOCK/125)))
+#define ticks_to_nanosecs(ticks)	((((u64)(ticks)*8000)/(u64)(TB_TIMER_CLOCK/125)))
 
-#define tick_microsecs(ticks)		((u32)(((u64)(ticks)*8)%(u64)(TB_TIMER_CLOCK/125)))
-#define tick_nanosecs(ticks)		((u32)(((u64)(ticks)*8000)%(u64)(TB_TIMER_CLOCK/125)))
+#define tick_microsecs(ticks)		((((u64)(ticks)*8)%(u64)(TB_TIMER_CLOCK/125)))
+#define tick_nanosecs(ticks)		((((u64)(ticks)*8000)%(u64)(TB_TIMER_CLOCK/125)))
 
 
 #define secs_to_ticks(sec)			((u64)(sec)*(TB_TIMER_CLOCK*1000))
